chore: Modularize fold
