+++ clar2wasm/src/wasm_generator.rs
@@ -291,16 +291,6 @@ impl WasmGenerator {
                 ) {
                     use clarity::vm::functions::NativeFunctions::*;
                     match native_function {
-                        Fold => {
-                            let name = args.get_name(0)?;
-                            self.traverse_fold(
-                                builder,
-                                expr,
-                                name,
-                                args.get_expr(1)?,
-                                args.get_expr(2)?,
-                            )
-                        }
                         FetchEntry => {
                             let name = args.get_name(0)?;
                             self.traverse_map_get(builder, expr, name, args.get_expr(1)?)
@@ -2480,141 +2470,6 @@ impl WasmGenerator {
                 .by_name(name.as_str())
                 .expect("function not found"),
         );
-        Ok(())
-    }
-
-    fn traverse_fold(
-        &mut self,
-        builder: &mut InstrSeqBuilder,
-        _expr: &SymbolicExpression,
-        func: &ClarityName,
-        sequence: &SymbolicExpression,
-        initial: &SymbolicExpression,
-    ) -> Result<(), GeneratorError> {
-        // Fold takes an initial value, and a sequence, and applies a function
-        // to the output of the previous call, or the initial value in the case
-        // of the first call, and each element of the sequence.
-        // ```
-        // (fold - (list 2 4 6) 0)
-        // ```
-        // is equivalent to
-        // ```
-        // (- 6 (- 4 (- 2 0)))
-        // ```
-
-        // The result type must match the type of the initial value
-        let result_clar_ty = self
-            .get_expr_type(initial)
-            .expect("fold's initial value expression must be typed");
-        let result_ty = clar2wasm_ty(result_clar_ty);
-        let loop_body_ty = InstrSeqType::new(&mut self.module.types, &[], &[]);
-
-        // Get the type of the sequence
-        let seq_ty = match self
-            .get_expr_type(sequence)
-            .expect("sequence expression must be typed")
-        {
-            TypeSignature::SequenceType(seq_ty) => seq_ty.clone(),
-            _ => {
-                return Err(GeneratorError::InternalError(
-                    "expected sequence type".to_string(),
-                ));
-            }
-        };
-
-        let (seq_len, elem_ty) = match &seq_ty {
-            SequenceSubtype::ListType(list_type) => {
-                (list_type.get_max_len(), list_type.get_list_item_type())
-            }
-            _ => unimplemented!("Unsupported sequence type"),
-        };
-
-        // Evaluate the sequence, which will load it into the call stack,
-        // leaving the offset and size on the data stack.
-        self.traverse_expr(builder, sequence)?;
-
-        // Drop the size, since we don't need it
-        builder.drop();
-
-        // Store the offset into a local
-        let offset = self.module.locals.add(ValType::I32);
-        builder.local_set(offset);
-
-        let elem_size = get_type_size(elem_ty);
-
-        // Store the end of the sequence into a local
-        let end_offset = self.module.locals.add(ValType::I32);
-        builder
-            .local_get(offset)
-            .i32_const(seq_len as i32 * elem_size)
-            .binop(BinaryOp::I32Add)
-            .local_set(end_offset);
-
-        // Evaluate the initial value, so that its result is on the data stack
-        self.traverse_expr(builder, initial)?;
-
-        if seq_len == 0 {
-            // If the sequence is empty, just return the initial value
-            return Ok(());
-        }
-
-        // Define local(s) to hold the intermediate result, and initialize them
-        // with the initial value. Not that we are looping in reverse order, to
-        // pop values from the top of the stack.
-        let mut result_locals = Vec::with_capacity(result_ty.len());
-        for local_ty in result_ty.iter().rev() {
-            let local = self.module.locals.add(*local_ty);
-            result_locals.push(local);
-            builder.local_set(local);
-        }
-        result_locals.reverse();
-
-        // Define the body of a loop, to loop over the sequence and make the
-        // function call.
-        builder.loop_(loop_body_ty, |loop_| {
-            let loop_id = loop_.id();
-
-            // Load the element from the sequence
-            let elem_size = self.read_from_memory(loop_, offset, 0, elem_ty);
-
-            // Push the locals to the stack
-            for result_local in result_locals.iter() {
-                loop_.local_get(*result_local);
-            }
-
-            // Call the function
-            loop_.call(
-                self.module
-                    .funcs
-                    .by_name(func.as_str())
-                    .expect("function not found"),
-            );
-
-            // Save the result into the locals (in reverse order as we pop)
-            for result_local in result_locals.iter().rev() {
-                loop_.local_set(*result_local);
-            }
-
-            // Increment the offset by the size of the element, leaving the
-            // offset on the top of the stack
-            loop_
-                .local_get(offset)
-                .i32_const(elem_size)
-                .binop(BinaryOp::I32Add)
-                .local_tee(offset);
-
-            // Loop if we haven't reached the end of the sequence
-            loop_
-                .local_get(end_offset)
-                .binop(BinaryOp::I32LtU)
-                .br_if(loop_id);
-        });
-
-        // Push the locals to the stack
-        for result_local in result_locals.iter() {
-            builder.local_get(*result_local);
-        }
-
         Ok(())
     }
